+++ src/Components/RenderStars.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const RenderStars = () => {
+  return <div></div>;
+};
+
+export default RenderStars;
