+++ src/lib/professionalItems.js
@@ -1,0 +1,27 @@
+export const professionalItems = [
+  "Agile and Digital Business Analysis",
+  "Product Management",
+  "Data Analytics",
+  "Power BI",
+  "Business Intelligence",
+  "Software Development",
+  "Digital Transformation Solutions",
+  "Cyber Security",
+  "Data Management",
+  "Data Strategy",
+  "Artificial Intelligence",
+  "Cloud Computing and more",
+];
+
+export const industriesItems = [
+  "Business Transformation & Advisory",
+  "Turnaround, Stability & Restructuring Transaction",
+  "Valuation & Financial Advisory Services",
+  "Dispute Advisory",
+  "Tax Services",
+  "Healthcare",
+  "Financial Services",
+  "Technology",
+  "Automotive",
+  "Oil & Gas",
+];
