+++ src/pages/auth/Signup.jsx
@@ -1,34 +1,31 @@
 import { zodResolver } from "@hookform/resolvers/zod";
 import { useState } from "react";
 import { useForm } from "react-hook-form";
-import { Link, useNavigate } from "react-router-dom";
+import { Link } from "react-router-dom";
 import { z } from "zod";
-import AviNav from "@/Components/avi/AviNav";
 import BorderCard from "@/Components/BorderCard";
 import { CommonButton } from "@/Components/ui/button";
 import { Form } from "@/Components/ui/form";
 import FormInput from "@/Components/ui/form-input";
-import { PasswordInput } from "@/Components/ui/password-input";
 import { Heading, Paragraph } from "./components/Text";
-import OtpComponent from "@/Components/about/OtpComponent";
 import Modal from "./components/Modal";
 import RegisterSuccess from "./components/RegisterSuccess";
+import ConfirmEmail from "./components/ConfirmEmail";
+import PasswordInput from "@/Components/ui/password-input";
 
 const loginSchema = z.object({
   email: z.string().email({ message: "Please enter a valid email address." }),
   password: z
     .string()
-    .min(4, { message: "Name must be at least 4 characters long" }),
-  referralCode: z
+    .min(4, { message: "Password must be at least 4 characters long " }),
+
+  confirmPassword: z
     .string()
     .min(4, { message: "Name must be at least 4 characters long" }),
-  connfirmPassword: z
-    .string()
-    .min(4, { message: "Name must be at least 4 characters long" }),
-  firstname: z
+  firstName: z
     .string()
     .min(1, { message: " first name must be at least 4 characters long" }),
-  lastname: z
+  lastName: z
     .string()
     .min(1, { message: "last name must be at least 4 characters long" }),
   username: z
@@ -37,62 +34,36 @@ const loginSchema = z.object({
 });
 
 const SignUp = () => {
-  // const [currentPassword, setCurrentPassword] = useState("");
-  const [password, setPassword] = useState("");
-  // const [passwordConfirmation, setPasswordConfirmation] = useState("");
   const [success, setSuccess] = useState("");
   const [confirm, setConfirm] = useState(false);
   const [modal, setModal] = useState(false);
-
-  const navigate = useNavigate();
 
   const form = useForm({
     resolver: zodResolver(loginSchema),
     defaultValues: {
       email: "",
       password: "",
-      firstname: "",
-      lastname: "",
+      firstName: "",
+      lastName: "",
       username: "",
-      connfirmPassword: "",
+      confirmPassword: "",
       referralCode: "",
     },
   });
+
+  const handleSubmit = (values) => {
+    console.log(values);
+  };
 
   return (
     <div className="min-h-screen">
       {confirm && (
         <Modal>
-          <BorderCard className="w-full max-w-[731px] rounded-xl bg-white py-11 text-center">
-            <div className="px-4">
-              <p className="text-xl font-semibold text-[#23314A]">
-                Confirm your email address
-              </p>
-              <p className="mx-auto mb-6 mt-3 max-w-[284px] text-center text-sm leading-[18px] text-[#98A2B3]">
-                Please enter code we sent now to aviplatform@gmail.com{" "}
-                <span className="text-primary-color-600">Edit</span>
-              </p>
-              <div className="mx-auto w-fit">
-                <OtpComponent />
-              </div>
-              <p className="mb-[31px] mt-6 text-sm">
-                <span className="text-[#645D5D]"> Didn’t receive a code?</span>{" "}
-                <span className="font-medium text-primary-color-600">
-                  Resend
-                </span>
-              </p>
-            </div>
-            <CommonButton
-              className="w-full bg-primary-color-600"
-              onClick={() => {
-                setConfirm((prev) => !prev);
-                setModal((prev) => !prev);
-                setSuccess("success");
-              }}
-            >
-              Confirm
-            </CommonButton>
-          </BorderCard>
+          <ConfirmEmail
+            setConfirm={setConfirm}
+            setModal={setModal}
+            setSuccess={setSuccess}
+          />
         </Modal>
       )}
       <div className="flex w-full items-center justify-center px-6 2xl:h-[calc(100vh-100.547px)]">
@@ -104,27 +75,22 @@ const SignUp = () => {
                 <Paragraph>Use your email to sign up</Paragraph>
               </div>
               <Form {...form}>
-                <form
-                  onSubmit={(e) => {
-                    e.preventDefault();
-                    setConfirm((prev) => !prev);
-                  }}
-                >
+                <form onSubmit={form.handleSubmit(handleSubmit)}>
                   <div className="space-y-[4px]">
                     <FormInput
                       label="firstname"
-                      name="firstname"
+                      name="firstName"
                       control={form.control}
                       type="text"
-                      id="firstname"
+                      id="firstName"
                       placeholder=""
                     />
                     <FormInput
                       label="lastname"
-                      name="lastname"
+                      name="lastName"
                       control={form.control}
                       type="text"
-                      id="lastname"
+                      id="lastName"
                       placeholder=""
                     />
                     <FormInput
@@ -145,8 +111,6 @@ const SignUp = () => {
                     />
                     <PasswordInput
                       id="password"
-                      value={password}
-                      onChange={(e) => setPassword(e.target.value)}
                       autoComplete="new-password"
                       label="password"
                       name="password"
@@ -155,8 +119,6 @@ const SignUp = () => {
                     />
                     <PasswordInput
                       id="confirmPassword"
-                      value={password}
-                      onChange={(e) => setPassword(e.target.value)}
                       autoComplete="new-password"
                       label="confirm password"
                       name="confirmPassword"
@@ -165,8 +127,6 @@ const SignUp = () => {
                     />
                     <PasswordInput
                       id="referralCode"
-                      value={password}
-                      onChange={(e) => setPassword(e.target.value)}
                       autoComplete="new-password"
                       label="referral Code"
                       name="referralCode"
