+++ src/pages/dashboard/DashboardDiscover.jsx
@@ -9,8 +9,8 @@ import { faStar, faTimes } from "@fortawesome/free-solid-svg-icons";
 import ModalContent from "../lms-pages/ReminderModalContent";
 import joinTeam from "../../assets/images/join_team.png";
 import styles from "../pages.module.css";
-import CourseCard, { DashboardDiscover } from "../../Components/CourseCard";
 import { Filter } from "@/Components/dashboard/Filter";
+import DashboardDiscover from "@/Components/DashboardDiscover";
 
 // import ReminderModal from '../auth/components/ReminderModal';
